- Generic way to filter by a field for tournaments - Changing prize from bigint to bigdec

+++ src/main/kotlin/com/hshar/eggegg/graphql/Mutation.kt
@@ -76,7 +76,7 @@ class Mutation : GraphQLMutationResolver {
         tournamentRepository.save(tournament)
 
         if (tournament.token.tokenVersion == 0) {
-            tournament.prize = Convert.fromWei(tournament.prize.toBigDecimal(), Convert.Unit.ETHER).toBigInteger()
+            tournament.prize = Convert.fromWei(tournament.prize, Convert.Unit.ETHER)
         }
         return tournament
     }
